Gitlab-Templatized/scripts/requirements.txt
pymssql==2.3.0
PyYAML==6.0.2
